change intro based on samuel's feedback
You have done an amazing job Nat, thank you . Just one thing at the moment, with the sentence:
"The Office for Disability Issues and Stats NZ have acknowledged that the cancellation should not have happened and have promised to consult more closely with the disability community in the future.".

As much as I wish they had said this, I think they were more saying it should not have been done without consultation (the cancellation may still have gone ahead, but they should have at least consulted the community).

So maybe change the sentence to:
"The Office for Disability Issues and Stats NZ have acknowledged that they should have consulted with the disability community and promise to do this in the future."